refactor(reactor): lift running a plan out of the command line
A run was never the command's: it is a configuration, a plan, a lock
declaration and a report, and none of that is about somebody standing at
a shell. It lived in `cli.py` all the same, so the only door to the
reactor was a subprocess, and anything else that wanted to run a plan
would have had to copy it.

The run moves to `reactorrun.py`, which takes the configuration it is to
run against instead of discovering one. `agentic-hil test-reactor` keeps
both entry points it had, unchanged in what they do and now the width of
their one contribution: which configuration a run is bound to. Nothing
else moved, so the plan loader, the workspace check, the lock
declaration, the per-step permissions and the report are the same lines
they were. The one test that reaches past the entry points to the
registered run follows it, patching the names where they now live.

Refs #231
